requirements.txt
Flask==1.1.1
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
oic==1.2.0
python-decouple==3.3
requests==2.22.0
python-dotenv==0.10.3
gunicorn==19.9.0
